+++ resmushit/resmushit.py
@@ -3,7 +3,7 @@ import os
 from urllib3 import PoolManager
 from .log import Log
 from .validator import Validator
-
+from typing import Union
 
 class Resmushit:
     """
@@ -31,11 +31,13 @@ class Resmushit:
     """
 
     __MAX_FILESIZE = 5 * 1024 * 1024
+    __API_URL: str = "http://api.resmush.it"
 
     def __init__(
         self,
         image_url: str = None,
         image_path: str = None,
+        _type: str = None,
         quality: int = 92,
         output_dir: str = ".",
         preserve_exif: bool = False,
@@ -51,43 +53,46 @@ class Resmushit:
             self.image_path = image_path
             self.image_url = None
         else:
-            raise ValueError("image_path or image_url is required")
-
+            raise ValueError("image_path or image_url param is required")
+        self._type = _type
         self.quality = quality
         self.output_dir = output_dir
         self.preserve_exif = preserve_exif
         self.preserve_filename = preserve_filename
         self.quiet_mode = quiet_mode
-        self.__API_URL: str = "http://api.resmush.it"
         self._response = None
         self.__logger = Log(quiet_mode=quiet_mode)
 
-    def __call_api(self):
+    @classmethod
+    def _filesize_limit(cls):
+        return cls.__MAX_FILESIZE
+
+    def __call_api(self) -> None:
         try:
             self.__logger.log(
-                message=f"Initializing image optimization with quality factor: {self.quality}%",
-                color="blue",
+                message=f"Initializing image optimization with quality factor: {self.quality}%"
             )
-            self.__logger.log(message=f"Sending picture {self.filename} to api...")
-
+            self.__logger.log(
+                message=f"Sending picture {self.filename}.{self.extension} to api..."
+            )
             http = PoolManager()
             r = http.request(
                 "POST",
-                self.__API_URL + f"/?qlty={self.quality}&exif={self.preserve_exif}",
-                fields={"files": ("image.png", self.imagebytes)},
+                Resmushit.__API_URL
+                + f"/?qlty={self.quality}&exif={self.preserve_exif}",
+                fields={"files": (f"{self.filename}.{self.extension}", self.imagebytes)},
             )
             self._response = json.loads(r.data.decode("utf-8"))
             self.__logger.log(
-                message=f"File optimized by {self._response.get('percent',0)}% (from {self._response.get('src_size',0)//1024}KB to {self._response.get('dest_size',0)//1024}KB). Retrieving...",
-                color="green",
+                message=f"File optimized by {self._response.get('percent',0)}% (from {self._response.get('src_size',0)//1024}KB to {self._response.get('dest_size',0)//1024}KB). Retrieving..."
             )
         except Exception as e:
             raise Exception(f"Error Occurred: {str(e)}")
 
-    def __get_dest_url(self):
+    def __get_dest_url(self) -> str:
         return self._response.get("dest")
 
-    def __download_image(self, dest_url):
+    def __download_image(self, dest_url) -> bytes:
         try:
             http = PoolManager()
             r = http.request("GET", dest_url)
@@ -95,24 +100,23 @@ class Resmushit:
         except Exception as e:
             raise Exception(f"{e}")
 
-    def __save_image(self, imagebytes):
+    def __save_image(self, imagebytes) -> bytes:
         with open(
             os.path.join(
                 self.output_dir,
                 f"{'' if self.preserve_filename else 'optimized-'}{self.filename}"
-                + f"{self.extension}",
+                +"."+ f"{self.extension}",
             ),
             "wb",
         ) as f:
             f.write(imagebytes)
 
-    def optimize(self, save: bool = True):
-        self.path, self.type, self.imagebytes, self.filename, self.extension = (
-            Validator(
-                path=self.image_url or self.image_path,
-                max_file_size=Resmushit.__MAX_FILESIZE,
-            ).validate()
-        )
+    def optimize(self, save: bool = True) -> Union[None, bytes]:
+        self.imagebytes, self.filename, self.extension = Validator(
+            path=self.image_url or self.image_path,
+            max_file_size=Resmushit.__MAX_FILESIZE,
+            _type=self._type,
+        ).validate()
         self.__logger.log(f"Processing: {self.filename}")
         self.__call_api()